Adding WebSockets test using WS scheme.
* WebSockets does not work after the merge of PR #358
* Adding an additional test using the ws scheme that was expected to work.
* Also updating tests in the same file to use the new test helper to cleanup communication objects.
